Restructure Student Analysis
