+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "module": "commonjs",
+    "module": "CommonJS",
     "moduleResolution": "node",
     "types": [
       "node"
@@ -12,7 +12,7 @@
     "rootDir": "./src",
     "outDir": "./lib",
     "baseUrl": "./src",
-    "removeComments": true
+    "removeComments": false
   },
   "exclude": [
     "node_modules"
